refactor tests, improve format
