+++ java/vector/src/test/java/org/apache/arrow/vector/file/TestArrowFile.java
@@ -27,55 +27,22 @@ import java.io.IOException;
 import java.util.List;
 
 import org.apache.arrow.memory.BufferAllocator;
-import org.apache.arrow.memory.RootAllocator;
 import org.apache.arrow.vector.FieldVector;
-import org.apache.arrow.vector.ValueVector.Accessor;
 import org.apache.arrow.vector.VectorLoader;
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.VectorUnloader;
 import org.apache.arrow.vector.complex.MapVector;
 import org.apache.arrow.vector.complex.NullableMapVector;
-import org.apache.arrow.vector.complex.impl.ComplexWriterImpl;
-import org.apache.arrow.vector.complex.reader.FieldReader;
-import org.apache.arrow.vector.complex.writer.BaseWriter.ComplexWriter;
-import org.apache.arrow.vector.complex.writer.BaseWriter.ListWriter;
-import org.apache.arrow.vector.complex.writer.BaseWriter.MapWriter;
-import org.apache.arrow.vector.complex.writer.BigIntWriter;
-import org.apache.arrow.vector.complex.writer.IntWriter;
-import org.apache.arrow.vector.file.json.JsonFileReader;
-import org.apache.arrow.vector.file.json.JsonFileWriter;
-import org.apache.arrow.vector.holders.NullableTimeStampHolder;
 import org.apache.arrow.vector.schema.ArrowBuffer;
 import org.apache.arrow.vector.schema.ArrowRecordBatch;
 import org.apache.arrow.vector.types.pojo.Schema;
-import org.joda.time.DateTimeZone;
-import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.netty.buffer.ArrowBuf;
-
-public class TestArrowFile {
+public class TestArrowFile extends BaseFileTest {
   private static final Logger LOGGER = LoggerFactory.getLogger(TestArrowFile.class);
-  private static final int COUNT = 10;
-  private BufferAllocator allocator;
-
-  private DateTimeZone defaultTimezone = DateTimeZone.getDefault();
-
-  @Before
-  public void init() {
-    DateTimeZone.setDefault(DateTimeZone.forOffsetHours(2));
-    allocator = new RootAllocator(Integer.MAX_VALUE);
-  }
-
-  @After
-  public void tearDown() {
-    allocator.close();
-    DateTimeZone.setDefault(defaultTimezone);
-  }
 
   @Test
   public void testWrite() throws IOException {
@@ -101,104 +68,6 @@ public class TestArrowFile {
       validateComplexContent(count, new VectorSchemaRoot(root));
       write(root, file);
     }
-  }
-
-  @Test
-  public void testWriteComplexJSON() throws IOException {
-    File file = new File("target/mytest_write_complex.json");
-    int count = COUNT;
-    try (
-        BufferAllocator vectorAllocator = allocator.newChildAllocator("original vectors", 0, Integer.MAX_VALUE);
-        NullableMapVector parent = new NullableMapVector("parent", vectorAllocator, null)) {
-      writeComplexData(count, parent);
-      VectorSchemaRoot root = new VectorSchemaRoot(parent.getChild("root"));
-      validateComplexContent(root.getRowCount(), root);
-      writeJSON(file, root);
-    }
-  }
-
-  public void writeJSON(File file, VectorSchemaRoot root) throws IOException {
-    JsonFileWriter writer = new JsonFileWriter(file, JsonFileWriter.config().pretty(true));
-    writer.start(root.getSchema());
-    writer.write(root);
-    writer.close();
-  }
-
-  @Test
-  public void testWriteReadComplexJSON() throws IOException {
-    File file = new File("target/mytest_complex.json");
-    int count = COUNT;
-
-    // write
-    try (
-        BufferAllocator originalVectorAllocator = allocator.newChildAllocator("original vectors", 0, Integer.MAX_VALUE);
-        MapVector parent = new MapVector("parent", originalVectorAllocator, null)) {
-      writeComplexData(count, parent);
-      writeJSON(file, new VectorSchemaRoot(parent.getChild("root")));
-    }
-
-    // read
-    try (
-        BufferAllocator readerAllocator = allocator.newChildAllocator("reader", 0, Integer.MAX_VALUE);
-        ) {
-      JsonFileReader reader = new JsonFileReader(file, readerAllocator);
-      Schema schema = reader.start();
-      LOGGER.debug("reading schema: " + schema);
-
-      // initialize vectors
-      try (VectorSchemaRoot root = reader.read();) {
-        validateComplexContent(count, root);
-      }
-      reader.close();
-    }
-  }
-
-  private void writeComplexData(int count, MapVector parent) {
-    ArrowBuf varchar = allocator.buffer(3);
-    varchar.readerIndex(0);
-    varchar.setByte(0, 'a');
-    varchar.setByte(1, 'b');
-    varchar.setByte(2, 'c');
-    varchar.writerIndex(3);
-    ComplexWriter writer = new ComplexWriterImpl("root", parent);
-    MapWriter rootWriter = writer.rootAsMap();
-    IntWriter intWriter = rootWriter.integer("int");
-    BigIntWriter bigIntWriter = rootWriter.bigInt("bigInt");
-    ListWriter listWriter = rootWriter.list("list");
-    MapWriter mapWriter = rootWriter.map("map");
-    for (int i = 0; i < count; i++) {
-      intWriter.setPosition(i);
-      intWriter.writeInt(i);
-      bigIntWriter.setPosition(i);
-      bigIntWriter.writeBigInt(i);
-      listWriter.setPosition(i);
-      listWriter.startList();
-      for (int j = 0; j < i % 3; j++) {
-        listWriter.varChar().writeVarChar(0, 3, varchar);
-      }
-      listWriter.endList();
-      mapWriter.setPosition(i);
-      mapWriter.start();
-      mapWriter.timeStamp("timestamp").writeTimeStamp(i);
-      mapWriter.end();
-    }
-    writer.setValueCount(count);
-    varchar.release();
-  }
-
-
-  private void writeData(int count, MapVector parent) {
-    ComplexWriter writer = new ComplexWriterImpl("root", parent);
-    MapWriter rootWriter = writer.rootAsMap();
-    IntWriter intWriter = rootWriter.integer("int");
-    BigIntWriter bigIntWriter = rootWriter.bigInt("bigInt");
-    for (int i = 0; i < count; i++) {
-      intWriter.setPosition(i);
-      intWriter.writeInt(i);
-      bigIntWriter.setPosition(i);
-      bigIntWriter.writeBigInt(i);
-    }
-    writer.setValueCount(count);
   }
 
   @Test
@@ -249,13 +118,6 @@ public class TestArrowFile {
     }
   }
 
-  private void validateContent(int count, VectorSchemaRoot root) {
-    for (int i = 0; i < count; i++) {
-      Assert.assertEquals(i, root.getVector("int").getAccessor().getObject(i));
-      Assert.assertEquals(Long.valueOf(i), root.getVector("bigInt").getAccessor().getObject(i));
-    }
-  }
-
   @Test
   public void testWriteReadComplex() throws IOException {
     File file = new File("target/mytest_complex.arrow");
@@ -293,45 +155,6 @@ public class TestArrowFile {
           validateComplexContent(count, root);
         }
       }
-    }
-  }
-
-  public void printVectors(List<FieldVector> vectors) {
-    for (FieldVector vector : vectors) {
-      LOGGER.debug(vector.getField().getName());
-      Accessor accessor = vector.getAccessor();
-      int valueCount = accessor.getValueCount();
-      for (int i = 0; i < valueCount; i++) {
-        LOGGER.debug(String.valueOf(accessor.getObject(i)));
-      }
-    }
-  }
-
-  private void validateComplexContent(int count, VectorSchemaRoot root) {
-    Assert.assertEquals(count, root.getRowCount());
-    printVectors(root.getFieldVectors());
-    for (int i = 0; i < count; i++) {
-      Assert.assertEquals(i, root.getVector("int").getAccessor().getObject(i));
-      Assert.assertEquals(Long.valueOf(i), root.getVector("bigInt").getAccessor().getObject(i));
-      Assert.assertEquals(i % 3, ((List<?>)root.getVector("list").getAccessor().getObject(i)).size());
-      NullableTimeStampHolder h = new NullableTimeStampHolder();
-      FieldReader mapReader = root.getVector("map").getReader();
-      mapReader.setPosition(i);
-      mapReader.reader("timestamp").read(h);
-      Assert.assertEquals(i, h.value);
-    }
-  }
-
-  private void write(FieldVector parent, File file) throws FileNotFoundException, IOException {
-    VectorUnloader vectorUnloader = newVectorUnloader(parent);
-    Schema schema = vectorUnloader.getSchema();
-    LOGGER.debug("writing schema: " + schema);
-    try (
-        FileOutputStream fileOutputStream = new FileOutputStream(file);
-        ArrowWriter arrowWriter = new ArrowWriter(fileOutputStream.getChannel(), schema);
-        ArrowRecordBatch recordBatch = vectorUnloader.getRecordBatch();
-            ) {
-      arrowWriter.writeRecordBatch(recordBatch);
     }
   }
 
@@ -433,69 +256,16 @@ public class TestArrowFile {
     }
   }
 
-  public void validateUnionData(int count, VectorSchemaRoot root) {
-    FieldReader unionReader = root.getVector("union").getReader();
-    for (int i = 0; i < count; i++) {
-      unionReader.setPosition(i);
-      switch (i % 4) {
-      case 0:
-        Assert.assertEquals(i, unionReader.readInteger().intValue());
-        break;
-      case 1:
-        Assert.assertEquals(i, unionReader.readLong().longValue());
-        break;
-      case 2:
-        Assert.assertEquals(i % 3, unionReader.size());
-        break;
-      case 3:
-        NullableTimeStampHolder h = new NullableTimeStampHolder();
-        unionReader.reader("timestamp").read(h);
-        Assert.assertEquals(i, h.value);
-        break;
-      }
+  private void write(FieldVector parent, File file) throws FileNotFoundException, IOException {
+    VectorUnloader vectorUnloader = newVectorUnloader(parent);
+    Schema schema = vectorUnloader.getSchema();
+    LOGGER.debug("writing schema: " + schema);
+    try (
+        FileOutputStream fileOutputStream = new FileOutputStream(file);
+        ArrowWriter arrowWriter = new ArrowWriter(fileOutputStream.getChannel(), schema);
+        ArrowRecordBatch recordBatch = vectorUnloader.getRecordBatch();
+            ) {
+      arrowWriter.writeRecordBatch(recordBatch);
     }
-  }
-
-  public void writeUnionData(int count, NullableMapVector parent) {
-    ArrowBuf varchar = allocator.buffer(3);
-    varchar.readerIndex(0);
-    varchar.setByte(0, 'a');
-    varchar.setByte(1, 'b');
-    varchar.setByte(2, 'c');
-    varchar.writerIndex(3);
-    ComplexWriter writer = new ComplexWriterImpl("root", parent);
-    MapWriter rootWriter = writer.rootAsMap();
-    IntWriter intWriter = rootWriter.integer("union");
-    BigIntWriter bigIntWriter = rootWriter.bigInt("union");
-    ListWriter listWriter = rootWriter.list("union");
-    MapWriter mapWriter = rootWriter.map("union");
-    for (int i = 0; i < count; i++) {
-      switch (i % 4) {
-      case 0:
-        intWriter.setPosition(i);
-        intWriter.writeInt(i);
-        break;
-      case 1:
-        bigIntWriter.setPosition(i);
-        bigIntWriter.writeBigInt(i);
-        break;
-      case 2:
-        listWriter.setPosition(i);
-        listWriter.startList();
-        for (int j = 0; j < i % 3; j++) {
-          listWriter.varChar().writeVarChar(0, 3, varchar);
-        }
-        listWriter.endList();
-        break;
-      case 3:
-        mapWriter.setPosition(i);
-        mapWriter.start();
-        mapWriter.timeStamp("timestamp").writeTimeStamp(i);
-        mapWriter.end();
-        break;
-      }
-    }
-    writer.setValueCount(count);
-    varchar.release();
   }
 }
